requirements.txt
accelerate==1.8.1
beartype==0.18.5
bitsandbytes==0.44.1
colorama==0.4.6
decord==0.6.0
easydict==1.13
gradio==4.44.0
jaraco.collections==5.1.0
kornia==0.7.3
modelscope==1.20.1
moviepy==1.0.3
numpy==1.26.0
omegaconf==2.3.0
open-clip-torch==2.0.2
openai==1.46.0
opencv-python==4.10.0.84
peft==0.16.0
pip-chill==1.0.3
prodigyopt==1.0
pytorch-lightning==2.3.3
scikit-learn==1.6.0
seaborn==0.13.2
swissarmytransformer==0.4.12
tensorboard==2.17.1
transformers>=4.46.2
tomli==2.0.1
torchao>=0.11.0
wandb==0.17.5
xformers==0.0.28.post1